vector_storage/weaviate_connector.py
import weaviate

class WeaviateConnector:
    def __init__(self, collection_name, embedding_model):
        self.client = weaviate.Client("http://localhost:8080")
        self.collection_name = collection_name
        self.embedding_model = embedding_model
    
    def is_new(self):
        return not self.client.schema.exists(self.collection_name)
    
    def store_documents(self, documents):
        # Store documents in Weaviate
        pass  # Implementation depends on data and schema structure
    
    def query_collection(self, query_text):
        # Query Weaviate collection to retrieve similar documents
        pass  # Add query implementation here