requirements.txt
pandas
numpy
weaviate-client
detoxify
torch
datasets
prometheus
logging